Motor: change sleep() to use stop() instead of brake()
brake() uses the short-brake function of the TB6612FNG, which causes a brief
short-circuit *across* the motor, causing it to act like a dynamo and it rapidly
loses kinetic energy. This short-circuit is disconnected from power, so it won't
cause damage other than a little heat while braking, which at this scale is
negligible.
It is possible that braking vs stopping will have a difference in precision, but
that is not of concern when simply putting the motor on standby.