README: Remove TODO on implementing joystick control - won't implement
Train of thought leading to this decision:
1. `pygame`, in the simplest way I could run it, has way too much latency and
has sometimes desyncs with large movements
2. Looked for a different library, and found `pyjoystick`. Looks nice, seems to
present a simple interface for reading joysticks, uses SDL directly but does all
the complex stuff of understanding SDL for me. Also very low latency.
3. `pyjoystick` implements its own event loop - not using asyncio or any other.
4. Digging through the `pyjoystick` library, the event loop is fundamental to
the whole library, making it difficult to just take a few functions into the
main event loop already in BioBox.
5. I *do not* want to mess with multiple event loops as that is sure to lead to
awkward issues with desyncs and deadlocks
6. SDL is sufficiently complex that it's not worth calling directly for this
part of a project.

Ultimately this was to be a "nice to have" fallback in case there is no slider,
but it would be about as involved to implement joystick support as it was to
originally implement the slider control.