VolSock: Discard yt_hook and just embed the extension ID in volsock.js
Turns out it's really hard for a page script to talk to an extension without
knowing who to talk to. There is no reason to keep the YouTube-specific
stuff in a separate file, nor is there a big problem hard-coding the
extension ID into volsock.js as that ID will become stable upon packing the
extension.