BioBox: Disable slider in GUI and don't start the task if not on a Pi
Previously the task would simply return immediately if Analog set the flag
no_slider, and this was sufficient. However, with the pending feature of
automatically restarting tasks that were not cancelled, the slider would
also automatically restart - this would not be harmful but for the console
spam. However, the disabling of the GUI elements makes it very clear that
the slider is not available.